pyproject.toml
[tool.poetry]
name = "class33"
version = "0.1.0"
description = ""
authors = ["majedalswaeer <majedalswaeer69@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
djangorestframework-jwt = "^1.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
